requirements.txt
streamlit
gspread
pandas
google-auth
sentence-transformers
torch